requirements.txt
Django==3.2.10
django-filter==21.1
django-extensions==3.1.5
django-celery-beat==2.2.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
celery==5.2.1
flower==1.0.0
MarkupSafe==2.0.1
pandas==1.3.5
psycopg2-binary==2.9.2
PyJWT==2.3.0
redis==4.0.2
requests==2.26.0
sentry-sdk==1.5.1
gevent==21.1.2
gunicorn
PyPDF2
python-pdf
pdf2docx
pillow==9.0.0

